Introducing Message to mark a field
The marked field can be marked (if the client recognizes the message)
Can be used in the production of video-blogs or strategy discussion, or
as a means to call Tak! on a specific field.

Markings can be set by players and Observers. Observers see all markings,
players only player markings. This prevents interference in the game.
A field can be marked by both player and observer. It is up to the client
to properly treat this state.
The marking of fields is game specific in preparation for players being
able to play multiple games simultaniously.